+++ packages/abi/lib/coders/array.d.ts
@@ -1,6 +1,8 @@
-import { Coder, Reader, Writer } from "./abstract-coder";
-export declare function pack(writer: Writer, coders: Array<Coder>, values: Array<any>): number;
-export declare function unpack(reader: Reader, coders: Array<Coder>): Array<any>;
+import { Coder, Reader, Result, Writer } from "./abstract-coder";
+export declare function pack(writer: Writer, coders: ReadonlyArray<Coder>, values: Array<any> | {
+    [name: string]: any;
+}): number;
+export declare function unpack(reader: Reader, coders: Array<Coder>): Result;
 export declare class ArrayCoder extends Coder {
     readonly coder: Coder;
     readonly length: number;