+++ packages/abstract-signer/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "abstract-signer/5.0.0-beta.143";
+export declare const version = "abstract-signer/5.0.0-beta.144";