+++ packages/abstract-signer/lib/index.js
@@ -237,6 +237,7 @@ var Signer = /** @class */ (function () {
                         if (tx.gasLimit == null) {
                             tx.gasLimit = this.estimateGas(tx).catch(function (error) {
                                 return logger.throwError("cannot estimate gas; transaction may fail or may require manual gas limit", logger_1.Logger.errors.UNPREDICTABLE_GAS_LIMIT, {
+                                    error: error,
                                     tx: tx
                                 });
                             });