+++ packages/tests/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "tests/5.0.0-beta.159";
+export declare const version = "tests/5.0.0-beta.160";