+++ packages/web/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "web/5.0.0-beta.139";
+export declare const version = "web/5.0.0-beta.140";