+++ packages/web/lib.esm/index.js
@@ -96,6 +96,7 @@ export function fetchJson(connection, json, processFunc) {
             catch (error) {
                 response = error.response;
                 if (response == null) {
+                    runningTimeout.cancel();
                     logger.throwError("missing response", Logger.errors.SERVER_ERROR, {
                         serverError: error,
                         url: url